+++ src/app/types/søkerinfo.ts
@@ -1,0 +1,9 @@
+import Person, { RegistrertAnnenForelder, RegistrertBarn } from './Person';
+import Arbeidsforhold from './Arbeidsforhold';
+
+export interface Søkerinfo {
+    person: Person;
+    arbeidsforhold?: Arbeidsforhold[];
+    registrertAnnenForelder?: RegistrertAnnenForelder;
+    registrerteBarn?: RegistrertBarn[];
+}
